requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via -r requirements.in, django
boto3==1.15.16            # via -r requirements.in
botocore==1.18.16         # via -r requirements.in, boto3, s3transfer
certifi==2020.6.20        # via -r requirements.in, requests
chardet==3.0.4            # via -r requirements.in, requests
django==3.1.2             # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via -r requirements.in, requests
jmespath==0.10.0          # via -r requirements.in, boto3, botocore
mysqlclient==2.0.1        # via -r requirements.in
pillow==7.2.0             # via -r requirements.in
python-dateutil==2.8.1    # via -r requirements.in, botocore
pytz==2020.1              # via -r requirements.in, django
requests==2.24.0          # via -r requirements.in
s3transfer==0.3.3         # via -r requirements.in, boto3
six==1.15.0               # via -r requirements.in, python-dateutil
sqlparse==0.3.1           # via -r requirements.in, django
urllib3==1.25.10          # via -r requirements.in, botocore, requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
